# src/swagger/swagger.yaml
basePath: /
definitions:
  api.BlockContactRequest:
    properties:
      wid:
        description: |-
          Wid is the WhatsApp ID (JID) of the contact to block or unblock.
          Example: "5511999999999@s.whatsapp.net"
        type: string
    type: object
...
  api.RabbitMQResponse:
    properties:
      affected:
        type: integer
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      rabbitmq:
        items:
          $ref: '#/definitions/models.QpRabbitMQConfig'
        type: array
      status:
        type: string
      success:
        type: boolean
    type: object
...
  models.QpRabbitMQConfig:
    properties:
      broadcasts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle broadcast messages
      calls:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle calls
      connection_string:
        description: RabbitMQ Connection Settings
        type: string
      deliveryreceipts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should emit delivery receipts
      direct:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle direct (individual) messages (@s.whatsapp.net and
          @lid); default true
      exchange_name:
        description: RabbitMQ exchange name for routing
        type: string
      extra:
        description: extra info to append on payload
      failure:
        description: Status Tracking
        type: string
      forwardinternal:
        description: Configuration Options
        type: boolean
      groups:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle groups messages
      queue_history:
        description: RabbitMQ history queue name (optional)
        type: string
      readreceipts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should emit read receipts
      readupdate:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should send markread requests when receiving messages
      routing_key:
        description: RabbitMQ routing key for exchange routing
        type: string
      success:
        description: last success timestamp
        type: string
      timestamp:
        type: string
      trackid:
        description: identifier of remote system to avoid loop
        type: string
      voipmode:
        allOf:
        - $ref: '#/definitions/whatsapp.VoIPMode'
        description: |-
          VoIPMode controls inbound WhatsApp call handling for this instance:
          disabled (default reject/relay), exclusive (SIP only, hang up WhatsApp on
          SIP failure), or additional (SIP as extra device, leave call ringing).
          Persisted in server metadata, not as a dedicated DB column.
    type: object
  models.QpResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
    type: object
...
  whatsapp.VoIPMode:
    enum:
    - disabled
    - exclusive
    - additional
    type: string
    x-enum-varnames:
    - VoIPModeDisabled
    - VoIPModeExclusive
    - VoIPModeAdditional
...
  whatsapp.WhatsappBoolean:
    enum:
    - -1
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - FalseBooleanType
    - UnSetBooleanType
    - TrueBooleanType
...
info:
  contact:
    name: QuePasa Support
    url: https://github.com/inteliagenciadigital/quepasa
  description: QuePasa is a Go-based WhatsApp bot platform that exposes HTTP APIs
    for WhatsApp messaging integration
  license:
    name: GNU Affero General Public License v3.0
    url: https://github.com/inteliagenciadigital/quepasa/blob/main/LICENSE.md
  termsOfService: https://github.com/inteliagenciadigital/quepasa
  title: QuePasa WhatsApp API
  version: 5.26.0625.0
paths:
  /account:
    post:
      consumes:
      - application/json
      description: Create, update, or manage user accounts (master access required)
      parameters:
      - description: Account request
        in: body
        name: request
        schema:
          properties:
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage user accounts
      tags:
      - Application
...
  /contacts/block:
    delete:
      consumes:
      - application/json
      description: Removes a previously placed block from a WhatsApp contact.
      parameters:
      - description: Contact to unblock
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.BlockContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Unblock contact
      tags:
      - Contacts
    post:
      consumes:
      - application/json
      description: Blocks a WhatsApp contact by their JID/WID so they cannot send
        messages to this session.
      parameters:
      - description: Contact to block
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.BlockContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Block contact
      tags:
      - Contacts
...
  /rabbitmq:
    delete:
      consumes:
      - application/json
      description: Create, get, or delete RabbitMQ configurations for message queueing
      parameters:
      - description: RabbitMQ config (for POST)
        in: body
        name: request
        schema:
          properties:
            connection_string:
              type: string
            exchange:
              type: string
            routing_key:
              type: string
          type: object
      - description: Connection string (for DELETE)
        in: query
        name: connection_string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RabbitMQResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage RabbitMQ configurations
      tags:
      - RabbitMQ
    get:
      consumes:
      - application/json
      description: Create, get, or delete RabbitMQ configurations for message queueing
      parameters:
      - description: RabbitMQ config (for POST)
        in: body
        name: request
        schema:
          properties:
            connection_string:
              type: string
            exchange:
              type: string
            routing_key:
              type: string
          type: object
      - description: Connection string (for DELETE)
        in: query
        name: connection_string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RabbitMQResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage RabbitMQ configurations
      tags:
      - RabbitMQ
    post:
      consumes:
      - application/json
      description: Create, get, or delete RabbitMQ configurations for message queueing
      parameters:
      - description: RabbitMQ config (for POST)
        in: body
        name: request
        schema:
          properties:
            connection_string:
              type: string
            exchange:
              type: string
            routing_key:
              type: string
          type: object
      - description: Connection string (for DELETE)
        in: query
        name: connection_string
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RabbitMQResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Manage RabbitMQ configurations
      tags:
      - RabbitMQ
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-QUEPASA-TOKEN
    type: apiKey
swagger: "2.0"